pyproject.toml
[project]
name = "hello-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.22",
]
